# Conflicts:
#	src/main/java/com/minecolonies/core/entity/ai/workers/production/EntityAIStructureMiner.java

+++ src/main/java/com/minecolonies/core/colony/jobs/JobHealer.java
@@ -71,7 +71,7 @@ public class JobHealer extends AbstractJob<EntityAIWorkHealer, JobHealer>
     }
 
     @Override
-    public int getDiseaseModifier()
+    public double getDiseaseModifier()
     {
         return 0;
     }
